requirements.txt
phidata
streamlit
openai
groq
sqlalchemy
pgvector
psycopg[binary]
pypdf